Revert "Move layer_num_blocks metric from syncer to tortoise"
This reverts commit 369ac83cb89fba5cd42a5bdf41ed96f58510aca4.
